+++ src/steam/steam_api.cpp
@@ -3083,7 +3083,7 @@ SK_Steam_ShouldThrottleCallbacks (void)
 
     if ( limit == 0 ||
       ( SK_CurrentPerf ().QuadPart - liLastCallbacked.QuadPart  <
-                        SK_PerfFreq / limit ) )
+              (LONGLONG)SK_PerfFreq / limit ) )
     {
       InterlockedDecrement64 ( &SK_SteamAPI_CallbackRunCount );
       return true;
